controller setup

+++ VirtualPetCareApi/VirtualPetCare.Core/Models/Activity.cs
@@ -1,6 +1,6 @@
 ﻿namespace VirtualPetCare.Core.Models;
 
-public class Activity : Entity<byte>
+public class Activity : Entity<int>
 {
     public List<Pet> Pets { get; set; }
 }
