+++ VirtualPetCareApi/VirtualPetCare.Core/Models/Food.cs
@@ -1,6 +1,6 @@
 ﻿namespace VirtualPetCare.Core.Models;
 
-public class Food : Entity<byte>
+public class Food : Entity<int>
 {
     public List<Pet> Pets { get; set; }
 }
