+++ VirtualPetCareApi/VirtualPetCare.Core/Models/HealthCondition.cs
@@ -1,6 +1,6 @@
 ﻿namespace VirtualPetCare.Core.Models;
 
-public class HealthCondition : Entity<byte>
+public class HealthCondition : Entity<int>
 {
     public int MyProperty { get; set; }
 }
